requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
click>=8.1.7
jinja2>=3.1.2
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=30.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
